src/main/java/org/fss/service/impl/UserServiceImpl.java
package org.fss.service.impl;

import org.fss.dto.UserDto;
import org.fss.entity.Role;
import org.fss.entity.User;
import org.fss.repository.RoleRepository;
import org.fss.repository.UserRepository;
import org.fss.service.UserService;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;
    private final RoleRepository roleRepository;
    private final PasswordEncoder passwordEncoder;

    public UserServiceImpl(UserRepository userRepository,
                           RoleRepository roleRepository,
                           PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.roleRepository = roleRepository;
        this.passwordEncoder = passwordEncoder;
    }

    //A function that save user and check for roles if role doesn't exist
    @Override
    public void saveUser(UserDto userDto) {
        User user = new User();
        user.setName(userDto.getFirstName() + " " + userDto.getLastName());
        user.setEmail(userDto.getEmail());
        // encrypt the password using spring security
        user.setPassword(passwordEncoder.encode(userDto.getPassword()));

        Role role = roleRepository.findByName("ROLE_USER");
        if (role == null) {
            role = checkRoleExist();
        }
        user.setRoles(List.of(role));


        userRepository.save(user);
    }

    //A function that finds user by email in database
    @Override
    public User findUserByEmail(String email) {
        return userRepository.findByEmail(email);
    }

    //A function that creates a role in database
    private Role checkRoleExist() {
        Role role = new Role();
        role.setName("ROLE_USER");
        return roleRepository.save(role);
    }
}
